Standardize all Provider JS files to be organized properly. Fix all deprecation warnings
