+++ src/js/features/classification/taxonomy-controls.js
@@ -278,6 +278,7 @@ const TaxonomyControls = ( { onChange, query } ) => {
 								value={ getExistingTaxQueryValue( slug ) }
 								suggestions={ terms.names }
 								onChange={ onTermsChange( slug ) }
+								__next40pxDefaultSize
 							/>
 							{ ! hasAI && (
 								<>
