+++ src/js/settings/components/feature-additional-settings/excerpt-generation.js
@@ -92,6 +92,7 @@ export const ExcerptGenerationSettings = () => {
 					onChange={ ( value ) =>
 						setFeatureSettings( { length: value } )
 					}
+					__next40pxDefaultSize
 				/>
 			</SettingsRow>
 		</>
