+++ src/js/settings/components/feature-additional-settings/nlu-feature.js
@@ -129,6 +129,7 @@ export const NLUFeatureSettings = () => {
 							min="0"
 							max="100"
 							step="0.01"
+							__next40pxDefaultSize
 						/>
 						{ 'ibm_watson_nlu' === featureSettings.provider && (
 							<SelectControl
@@ -155,6 +156,7 @@ export const NLUFeatureSettings = () => {
 									} );
 								} }
 								__nextHasNoMarginBottom
+								__next40pxDefaultSize
 							/>
 						) }
 					</SettingsRow>
