+++ src/js/settings/components/feature-additional-settings/term-cleanup.js
@@ -141,6 +141,7 @@ export const TermCleanupSettings = () => {
 								min="0"
 								max="100"
 								step="0.01"
+								__next40pxDefaultSize
 							/>
 						</SettingsRow>
 					);
