+++ src/js/settings/components/model-selector/index.js
@@ -67,6 +67,7 @@ export const ModelSelector = ( {
 				options={ models }
 				disabled={ models.length <= 1 }
 				__nextHasNoMarginBottom
+				__next40pxDefaultSize
 			/>
 		</SettingsRow>
 	);
