+++ src/js/settings/components/provider-settings/azure/azure-openai/openai.js
@@ -64,6 +64,7 @@ export const AzureOpenAISettings = ( {
 						onChange={ ( value ) =>
 							onChange( { number_of_suggestions: value } )
 						}
+						__next40pxDefaultSize
 					/>
 				</SettingsRow>
 			) }
