--- src/js/settings/components/provider-settings/ollama-embeddings.js
@@ -1,43 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { useSelect, useDispatch } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
-import { STORE_NAME } from '../../data/store';
-import { OllamaBaseSettings } from './ollama-base';
-
-/**
- * React Component for Ollama Embeddings settings.
- *
- * This component is used within the ProviderSettings component to
- * allow users to configure the Ollama Embeddings settings.
- *
- * @param {Object}  props              Component props.
- * @param {boolean} props.isConfigured Whether the provider is configured.
- *
- * @return {React.ReactElement} OllamaEmbeddingsSettings component.
- */
-export const OllamaEmbeddingsSettings = ( { isConfigured = false } ) => {
-	const providerName = 'ollama_embeddings';
-	const providerSettings = useSelect(
-		( select ) =>
-			select( STORE_NAME ).getFeatureSettings( providerName ) || {}
-	);
-	const { setProviderSettings } = useDispatch( STORE_NAME );
-	const onChange = ( data ) => setProviderSettings( providerName, data );
-
-	if ( isConfigured ) {
-		return null;
-	}
-
-	return (
-		<OllamaBaseSettings
-			providerSettings={ providerSettings }
-			providerName={ providerName }
-			onChange={ onChange }
-		/>
-	);
-};
