+++ src/js/settings/components/provider-settings/openai/text-to-speech.js
@@ -6,20 +6,18 @@ import {
 	SelectControl,
 	TextareaControl,
 } from '@wordpress/components';
+import { useSelect, useDispatch } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { SettingsRow } from '../settings-row';
-import { useSelect, useDispatch } from '@wordpress/data';
-import { STORE_NAME } from '../../data/store';
-import { OpenAISettings } from './openai';
+import { SettingsRow } from '../../settings-row';
+import { STORE_NAME } from '../../../data/store';
+import { OpenAIBaseSettings } from './base';
 
 /**
  * Component for OpenAI Text to Speech Provider settings.
- *
- * This component is used within the ProviderSettings component to allow users to configure the OpenAI Text to Speech Provider settings.
  *
  * @param {Object}  props              Component props.
  * @param {boolean} props.isConfigured Whether the provider is configured.
@@ -38,7 +36,7 @@ export const OpenAITextToSpeechSettings = ( { isConfigured = false } ) => {
 	return (
 		<>
 			{ ! isConfigured && (
-				<OpenAISettings
+				<OpenAIBaseSettings
 					providerSettings={ providerSettings }
 					onChange={ onChange }
 				/>
